Migrate to composite builds from buildSrc
Also
- Bump AGP version to 7.4.2 from 7.3.1
- Bump Kotlin version to 1.8.10 from 1.7.10
- Use KSP instead of kapt for Room
