+++ app/build.gradle.kts
@@ -1,81 +1,8 @@
-import extensions.buildConfigStringField
-import extensions.getLocalProperty
-
 plugins {
-    id(BuildPlugins.ANDROID_APPLICATION)
-    id(BuildPlugins.KOTLIN_KAPT)
-    id(BuildPlugins.KOTLIN_ANDROID)
-    id(BuildPlugins.HILT_PLUGIN)
-    id(BuildPlugins.NAVIGATION_SAFEARGS)
-    id(BuildPlugins.KOTLINX_SERIALIZATION)
-    id(BuildPlugins.JACOCO)
-}
-
-android {
-    compileSdk = AndroidConfig.COMPILE_SDK
-
-    defaultConfig {
-        applicationId = AndroidConfig.APP_ID
-        minSdk = AndroidConfig.MIN_SDK
-        targetSdk = AndroidConfig.TARGET_SDK
-        versionCode = AndroidConfig.VERSION_CODE
-        versionName = AndroidConfig.VERSION_NAME
-        testInstrumentationRunner = AndroidConfig.TEST_INSTRUMENTATION_RUNNER
-    }
-    signingConfigs {
-        BuildDebugSigningConfig.create(file("../debug.keystore"), this)
-        BuildReleaseSigningConfig.create(file("../release.keystore"), this)
-    }
-    buildTypes {
-        BuildTypeDebug.create(this)
-        BuildTypeRelease.create(this)
-        BuildTypeBenchmark.create(this, signingConfigs)
-    }
-    buildTypes.forEach {
-        kotlin.runCatching {
-            it.buildConfigStringField(BuildConstants.CONFIG_NAME_API_BASE_URL, BuildConstants.CONFIG_VALUE_API_BASE_URL)
-            it.buildConfigStringField(
-                BuildConstants.CONFIG_NAME_API_KEY,
-                getLocalProperty(BuildConstants.PROPERTY_NAME_API_KEY),
-            )
-        }.onFailure {
-            throw InvalidUserDataException(BuildConstants.MESSAGE_API_KEY_EXCEPTION)
-        }
-    }
-    flavorDimensions.add(BuildProductDimensions.DIMENSION_VERSION)
-    productFlavors {
-        ProductFlavorDev.create(this, true)
-        ProductFlavorQA.create(this, true)
-        ProductFlavorProduction.create(this, true)
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_1_8.toString()
-        freeCompilerArgs = freeCompilerArgs
-            .plus("-Xopt-in=kotlin.RequiresOptIn")
-            .plus("-Xcontext-receivers")
-    }
-    buildFeatures {
-        viewBinding = true
-    }
-    packagingOptions {
-        resources.excludes.add("META-INF/AL2.0")
-        resources.excludes.add("META-INF/LGPL2.1")
-    }
-    testOptions {
-        managedDevices {
-            devices {
-                create<com.android.build.api.dsl.ManagedVirtualDevice>("pixel6Api31") {
-                    device = "Pixel 6"
-                    apiLevel = 31
-                }
-            }
-        }
-    }
-    namespace = "com.azizutku.movie"
+    id("movie.android.application")
+    id("movie.android.hilt")
+    id("movie.android.room")
+    id("movie.android.application.jacoco")
 }
 
 dependencies {
@@ -89,7 +16,6 @@ dependencies {
     implementation(libs.androidx.swiperefreshlayout)
     implementation(libs.bundles.androidx.navigation)
     implementation(libs.bundles.androidx.lifecycle)
-    implementation(libs.bundles.androidx.room)
     implementation(libs.material)
     implementation(libs.hilt)
     implementation(libs.coroutines)
@@ -120,11 +46,6 @@ dependencies {
     androidTestImplementation(libs.navigation.androidTest)
     androidTestImplementation(libs.coroutines.test)
     androidTestImplementation(libs.arch.core.test)
-
-    kapt(libs.hilt.kapt)
-    kapt(libs.room.kapt)
-    kaptAndroidTest(libs.hilt.kapt)
-    kaptTest(libs.hilt.kapt)
 }
 
 kapt {
