+++ build-logic/convention/src/main/kotlin/com/azizutku/movie/AndroidConfig.kt
@@ -1,0 +1,15 @@
+package com.azizutku.movie
+
+object AndroidConfig {
+
+    const val APP_ID = "com.azizutku.movie"
+    const val COMPILE_SDK = 33
+    const val MIN_SDK = 23
+    const val TARGET_SDK = 33
+    const val VERSION_CODE = 1
+    const val VERSION_NAME = "1.0"
+    const val TEST_INSTRUMENTATION_RUNNER = "com.azizutku.movie.common.di.HiltTestRunner"
+    const val BENCHMARK_TARGET_PROJECT_PATH = ":app"
+    const val NAMESPACE = "com.azizutku.movie"
+
+}
