+++ build-logic/convention/src/main/kotlin/com/azizutku/movie/BuildConstants.kt
@@ -1,0 +1,12 @@
+package com.azizutku.movie
+
+object BuildConstants {
+
+    const val CONFIG_NAME_API_BASE_URL = "TMDB_API_BASE_URL"
+    const val CONFIG_VALUE_API_BASE_URL = "https://api.themoviedb.org/3/"
+    const val CONFIG_NAME_API_KEY = "TMDB_API_KEY"
+    const val MESSAGE_API_KEY_EXCEPTION = "You should define 'tmdb.api.key' in local.properties. " +
+        "You can get api key from https://www.themoviedb.org/settings/api"
+    const val PROPERTY_NAME_API_KEY = "tmdb.api.key"
+
+}
