+++ build-logic/convention/src/main/kotlin/com/azizutku/movie/BuildPlugins.kt
@@ -1,0 +1,13 @@
+package com.azizutku.movie
+
+object BuildPlugins {
+
+    const val ANDROID_APPLICATION = "com.android.application"
+    const val KOTLIN_ANDROID = "kotlin-android"
+    const val KOTLIN_KAPT = "kotlin-kapt"
+    const val HILT_PLUGIN = "dagger.hilt.android.plugin"
+    const val KOTLINX_SERIALIZATION = "kotlinx-serialization"
+    const val NAVIGATION_SAFEARGS = "androidx.navigation.safeargs.kotlin"
+    const val ANDROID_TEST = "com.android.test"
+    const val GRADLE_JACOCO = "org.gradle.jacoco"
+}
