+++ build-logic/convention/src/main/kotlin/com/azizutku/movie/BuildProductDimensions.kt
@@ -1,0 +1,5 @@
+package com.azizutku.movie
+
+object BuildProductDimensions {
+    const val DIMENSION_VERSION = "version"
+}
