+++ build-logic/convention/src/main/kotlin/com/azizutku/movie/extensions/BuildTypeExtensions.kt
@@ -1,0 +1,7 @@
+package com.azizutku.movie.extensions
+
+import com.android.build.api.dsl.ApplicationBuildType
+
+fun ApplicationBuildType.buildConfigStringField(name: String, value: String) {
+    this.buildConfigField("String", name, "\"$value\"")
+}
