+++ build-logic/convention/src/main/kotlin/com/azizutku/movie/utils/BuildUtils.kt
@@ -1,0 +1,19 @@
+package com.azizutku.movie.utils
+
+import org.gradle.api.Project
+import java.util.Properties
+
+private const val LOCAL_PROPERTIES_FILE_NAME = "local.properties"
+
+fun getLocalProperty(propertyName: String, project: Project): String {
+    val localProperties = Properties().apply {
+        val localPropertiesFile = project.rootProject.file(LOCAL_PROPERTIES_FILE_NAME)
+        if (localPropertiesFile.exists()) {
+            load(localPropertiesFile.inputStream())
+        }
+    }
+
+    return localProperties.getProperty(propertyName) ?: run {
+        throw NoSuchFieldException("Not defined property: $propertyName")
+    }
+}
