+++ build-logic/gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,3 @@
-#Sun Sep 18 13:12:04 TRT 2022
 distributionBase=GRADLE_USER_HOME
 distributionUrl=https\://services.gradle.org/distributions/gradle-7.6-bin.zip
 distributionPath=wrapper/dists
