+++ build.gradle.kts
@@ -1,18 +1,21 @@
+import com.azizutku.movie.utils.VersionUtils.isNonStable
 import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
-import utils.VersionUtils.isNonStable
 
 plugins {
     alias(libs.plugins.ktlint)
     alias(libs.plugins.versions)
     alias(libs.plugins.hilt).apply(false)
     alias(libs.plugins.kotlin.serialization).apply(false)
+    alias(libs.plugins.ksp).apply(false)
+    id("movie.git.hooks").apply(false)
+    id("movie.detekt").apply(false)
 }
 
-plugins.apply(BuildPlugins.GIT_HOOKS)
+plugins.apply("movie.git.hooks")
 
 allprojects {
-    plugins.apply(BuildPlugins.DETEKT)
-    plugins.apply(BuildPlugins.KTLINT_PLUGIN)
+    plugins.apply("movie.detekt")
+    plugins.apply("org.jlleitschuh.gradle.ktlint")
 }
 
 tasks.register("clean", Delete::class) {
