--- buildSrc/.gitignore
@@ -1,2 +1,0 @@
-/build
-/.gradle
