--- buildSrc/src/main/kotlin/BuildPlugins.kt
@@ -1,15 +1,0 @@
-object BuildPlugins {
-
-    const val ANDROID_APPLICATION = "com.android.application"
-    const val KOTLIN_ANDROID = "kotlin-android"
-    const val KOTLIN_KAPT = "kotlin-kapt"
-    const val KTLINT_PLUGIN = "org.jlleitschuh.gradle.ktlint"
-    const val HILT_PLUGIN = "dagger.hilt.android.plugin"
-    const val KOTLINX_SERIALIZATION = "kotlinx-serialization"
-    const val NAVIGATION_SAFEARGS = "androidx.navigation.safeargs.kotlin"
-    const val ANDROID_TEST = "com.android.test"
-
-    const val DETEKT = "plugins.detekt"
-    const val GIT_HOOKS = "plugins.git-hooks"
-    const val JACOCO = "plugins.jacoco"
-}
