--- buildSrc/src/main/kotlin/BuildProductDimensions.kt
@@ -1,3 +1,0 @@
-object BuildProductDimensions {
-    const val DIMENSION_VERSION = "version"
-}
