--- buildSrc/src/main/kotlin/extensions/BuildTypeExtensions.kt
@@ -1,7 +1,0 @@
-package extensions
-
-import com.android.build.gradle.internal.dsl.BuildType
-
-fun BuildType.buildConfigStringField(name: String, value: String) {
-    this.buildConfigField("String", name, "\"$value\"")
-}
