--- buildSrc/src/main/kotlin/extensions/ProjectExtensions.kt
@@ -1,8 +1,0 @@
-package extensions
-
-import org.gradle.api.Project
-import utils.getLocalProperty
-
-fun Project.getLocalProperty(propertyName: String): String {
-    return getLocalProperty(propertyName, this)
-}
