--- buildSrc/src/main/kotlin/utils/BuildUtils.kt
@@ -1,19 +1,0 @@
-package utils
-
-import org.gradle.api.Project
-import java.util.Properties
-
-private const val LOCAL_PROPERTIES_FILE_NAME = "local.properties"
-
-fun getLocalProperty(propertyName: String, project: Project): String {
-    val localProperties = Properties().apply {
-        val localPropertiesFile = project.rootProject.file(LOCAL_PROPERTIES_FILE_NAME)
-        if (localPropertiesFile.exists()) {
-            load(localPropertiesFile.inputStream())
-        }
-    }
-
-    return localProperties.getProperty(propertyName) ?: run {
-        throw NoSuchFieldException("Not defined property: $propertyName")
-    }
-}
