+++ gradle/libs.versions.toml
@@ -33,11 +33,12 @@ turbine = "0.12.1"
 archCore = "2.1.0"
 
 # Dependency versions of the included build-logic
-androidGradlePlugin = "7.3.1"
-kotlin = "1.7.10"
+androidGradlePlugin = "7.4.2"
+kotlin = "1.8.10"
 detekt = "1.19.0"
 javapoet = "1.13.0"
 navigationSafeargs = "2.5.3"
+kspGradlePlugin = "1.8.10-1.0.9"
 
 # Plugins
 ktlintPlugin = "11.0.0"
@@ -110,8 +111,8 @@ uiautomator-androidTest = { group = "androidx.test.uiautomator", name = "uiautom
 benchmark-macro-androidTest = { group = "androidx.benchmark", name = "benchmark-macro-junit4", version.ref = "benchmarkMacro" }
 
 # Annotation processor dependencies
-hilt-kapt = { group = "com.google.dagger", name = "hilt-compiler", version.ref = "hilt" }
-room-kapt = { group = "androidx.room", name = "room-compiler", version.ref = "room" }
+hilt-compiler = { group = "com.google.dagger", name = "hilt-compiler", version.ref = "hilt" }
+room-compiler = { group = "androidx.room", name = "room-compiler", version.ref = "room" }
 
 # Dependencies of the included build-logic
 android-gradle-plugin = { group = "com.android.tools.build", name = "gradle", version.ref = "androidGradlePlugin" }
@@ -119,6 +120,7 @@ kotlin-plugin = { group = "org.jetbrains.kotlin", name = "kotlin-gradle-plugin",
 detekt-plugin = { group = "io.gitlab.arturbosch.detekt", name = "detekt-gradle-plugin", version.ref = "detekt" }
 javapoet-plugin = { module = "com.squareup:javapoet", version.ref = "javapoet" }
 navigation-safeargs-plugin = { group = "androidx.navigation", name = "navigation-safe-args-gradle-plugin", version.ref = "navigationSafeargs" }
+ksp-gradle-plugin = { group = "com.google.devtools.ksp", name = "com.google.devtools.ksp.gradle.plugin", version.ref = "kspGradlePlugin" }
 
 [bundles]
 androidx-navigation = ["androidx.navigation.fragment", "androidx.navigation.ui"]
@@ -130,3 +132,4 @@ ktlint = { id = "org.jlleitschuh.gradle.ktlint", version.ref = "ktlintPlugin" }
 versions = { id = "com.github.ben-manes.versions", version.ref = "versionsPlugin" }
 hilt = { id = "dagger.hilt.android.plugin", version.ref = "hiltPlugin" }
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
+ksp = { id = "com.google.devtools.ksp", version.ref = "kspGradlePlugin" }
